+++ package.json
@@ -19,6 +19,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-icons": "^5.4.0",
+    "react-router-dom": "^7.2.0",
     "react-scripts": "^5.0.1",
     "resolve-url-loader": "^5.0.0",
     "rollup-plugin-terser": "^7.0.2",
