+++ src/App.js
@@ -1,35 +1,30 @@
 import React, { useState } from "react";
-import LoginForm from './components/LoginForm';
-import RegisterForm from './components/RegisterForm';
-import Header from './components/Header.js';
-import Student from "./models/Student.js";
 import StudentHome from "./components/StudentHome.js";
+import FeedbackForm from "./components/FeedbackForm.js";
+import Home from "./components/Home.js";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import Header from "./components/Header.js";
+
+
 
 const App = () => {
-  const [isSignup, setIsSignup] = useState(false);
-  //avoid having header on signup page
-  const [isLoginPage, setIsLoginPage] = useState(true);
-  const handleFormSwitch = () => {
-    setIsSignup(!isSignup);
-    setIsLoginPage(false);
-  };
-
   return (
     <div>
     {/* Display header only if not on login or signup page*/}
     {/*!isLoginPage && !isSignup && <Header/>*/} 
-       <Header/>
-      
-      
+    <Header/>   
+   
+   {/* Définition des différentes routes */}
+    <Router>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/feedback" element={<FeedbackForm />} />
+        <Route path="/studenthome" element={<StudentHome />} />
+      </Routes>
+    </Router>
 
-      {isSignup ? (
-        <RegisterForm onSwitch={handleFormSwitch} />
-      ) : isLoginPage ?(
-        <LoginForm onSwitch={handleFormSwitch} />
-      ):(
-        Student && <StudentHome studentName={Student.name} studentPic={Student.pic} />
-      )}
     </div>
+
   );
 };
