+++ src/components/FeedbackForm.js
@@ -1,0 +1,5 @@
+function FeedbackForm() {
+    return <h2>Hi !</h2>;
+  }
+
+  export default FeedbackForm
